adding strategies to build a multi tenant saas application for users
