--- src/db-utils/qdrant-adapter.service.ts/qdrant-adapter.service.ts.service.ts
@@ -1,17 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import {
-  EmbedDocumentsOptions,
-  VectorDBClient,
-} from '../vector-db-client.interface';
-@Injectable()
-export class QdrantAdapterService implements VectorDBClient {
-  delete(indexName: string): Promise<void> {
-    throw new Error('Method not implemented.');
-  }
-  describeIndexStats(indexName: string): Promise<any> {
-    throw new Error('Method not implemented.');
-  }
-  embedDocuments(options: EmbedDocumentsOptions): Promise<void> {
-    throw new Error('Method not implemented.');
-  }
-}
